requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
django-cors-headers==4.2.0
langchain===0.0.310
pypdf=3.16.3
rapidocr-onnxruntime==1.3.7
python-dotenv==1.0.0
openai==0.28.1
deeplake==3.7.3
tiktoken==0.5.1

